aider: Set eDP-1 monitor as the main screen and adjust desktop numbers accordingly.
# Aider chat conversation:

USER: it is seperated already, but can I make my monitor as main screen

ASSISTANT: Certainly! To set your eDP-1 monitor as the main screen, you will need to adjust the `bspwmrc` configuration file. This involves changing the order in which the monitors are set up and designating the eDP-1 monitor as the primary one.

Here's what you need to do in the `bspwmrc` file:

```bash
bspwm/bspwmrc
<<<<<<< SEARCH
 if xrandr | grep "HDMI-1-0 connected"; then
     xrandr --output HDMI-1-0 --right-of eDP-1 --mode 1920x1080 --rate 60.00
     bspc monitor HDMI-1-0 -d 1 2 3 4 5
     bspc monitor eDP-1 -d 6 7 8 9 0
 else
     bspc monitor eDP-1 -d 1 2 3 4 5 6 7 8 9 0
 fi
=======
 if xrandr | grep "HDMI-1-0 connected"; then
     xrandr --output eDP-1 --primary --mode 1920x1080 --rate 60.00 --output HDMI-1-0 --right-of eDP-1 --mode 1920x1080 --rate 60.00
     bspc monitor eDP-1 -d 1 2 3 4 5
     bspc monitor HDMI-1-0 -d 6 7 8 9 0
 else
     xrandr --output eDP-1 --primary --mode 1920x1080 --rate 60.00
     bspc monitor eDP-1 -d 1 2 3 4 5 6 7 8 9 0
 fi
>>>>>>> REPLACE
```

In this change, we've added the `--primary` flag to the `eDP-1` monitor configuration to set it as the main display. Additionally, we've swapped the desktop numbers so that the first five desktops are on the eDP-1 monitor, and the remaining five are on the HDMI-1-0 monitor when it is connected. If the HDMI-1-0 monitor is not connected, all ten desktops will be on the eDP-1 monitor.