aider: Launch both main and `bottom` configurations for each monitor.
# Aider chat conversation:

USER: each monitor should lauch both main and `bottom` configuratio

ASSISTANT: To launch Polybar with both the `main` and `bottom` configurations for each connected monitor, we need to modify the `polybar/launch.sh` script. We will add a second Polybar instance for each monitor with the `bottom` configuration.

Here's the change that needs to be made to `polybar/launch.sh`:

```sh
polybar/launch.sh
<<<<<<< SEARCH
  for m in $(xrandr --query | grep " connected" | cut -d" " -f1); do
    MONITOR=$m polybar --reload -c ~/.config/polybar/config_bottom.ini main &
  done
=======
  for m in $(xrandr --query | grep " connected" | cut -d" " -f1); do
    # Launch the main bar
    MONITOR=$m polybar --reload -c ~/.config/polybar/config.ini main &
    # Launch the bottom bar
    MONITOR=$m polybar --reload -c ~/.config/polybar/config_bottom.ini bottom &
  done
>>>>>>> REPLACE
```

This modification will start two instances of Polybar for each connected monitor: one with the `main` configuration and one with the `bottom` configuration. The `main` configuration is assumed to be the top bar, and the `bottom` configuration is for the bottom bar. Make sure that the `config.ini` and `config_bottom.ini` files are correctly configured for the top and bottom bars, respectively.