aider: Launch the bottom Polybar only on the HDMI-1-0 monitor when it's connected.
# Aider chat conversation:

ASSISTANT: Great, now that you've provided the `polybar/launch.sh` file, we can modify it to ensure that the bottom Polybar starts on the HDMI-1-0 monitor when it's connected.

Here's what you need to change in the `polybar/launch.sh` file:

```bash
polybar/launch.sh
<<<<<<< SEARCH
# Launch Polybar for each connected monitor, with special handling for eDP-1
# Launch Polybar with config_bottom.ini for all connected monitors
if type "xrandr"; then
  for m in $(xrandr --query | grep " connected" | cut -d" " -f1); do
    # Launch the main bar
    MONITOR=$m polybar --reload -c ~/.config/polybar/config.ini main &
    # Launch the bottom bar
    MONITOR=$m polybar --reload -c ~/.config/polybar/config_bottom.ini main &
  done
else
=======
# Launch Polybar for each connected monitor, with special handling for HDMI-1-0
# Launch Polybar with config_bottom.ini for HDMI-1-0 and config.ini for others
if type "xrandr"; then
  for m in $(xrandr --query | grep " connected" | cut -d" " -f1); do
    # Launch the main bar
    MONITOR=$m polybar --reload -c ~/.config/polybar/config.ini main &
    # Launch the bottom bar only on HDMI-1-0
    if [ "$m" = "HDMI-1-0" ]; then
      MONITOR=$m polybar --reload -c ~/.config/polybar/config_bottom.ini bottom &
    fi
  done
else
>>>>>>> REPLACE
```

This change will launch the bottom Polybar only on the HDMI-1-0 monitor when it's connected. If HDMI-1-0 is not connected, only the main Polybar will be launched on the available monitors. Make sure to replace `bottom` with the actual name of your bottom bar in the Polybar configuration if it's different.